Deduplicated reporting methods in base page and mobile base page
