+++ page_objects/Base_Page.py
@@ -10,6 +10,7 @@ import time,logging,os,inspect
 from utils.Base_Logging import Base_Logging
 from .driverfactory import DriverFactory
 from .selenium_objects import Selenium_Objects
+from .test_reporting_objects import Test_Reporting_Objects
 from .logging_objects import Logging_Objects
 from page_objects import PageFactory
 from utils.stop_test_exception_util import Stop_Test_Exception
@@ -38,7 +39,7 @@ class Borg:
 # Get the Base URL from the conf file
 base_url = conf.base_url_conf
 
-class Base_Page(Borg, Selenium_Objects, Logging_Objects):
+class Base_Page(Borg, Selenium_Objects, Logging_Objects, Test_Reporting_Objects):
     "Page class that all page models can inherit from"
 
     def __init__(self,base_url):
@@ -110,33 +111,9 @@ class Base_Page(Borg, Selenium_Objects, Logging_Objects):
 
         self.start()
 
-
     def get_current_driver(self):
         "Return current driver."
         return self.driver
-
-
-    def register_testrail(self):
-        "Register TestRail with Page"
-        self.testrail_flag = True
-        from utils.Test_Rail import Test_Rail  # pylint: disable=import-error,import-outside-toplevel
-        self.testrail_object = Test_Rail()
-
-    def set_test_run_id(self,test_run_id):
-        "Set TestRail's test run id"
-        self.test_run_id = test_run_id
-
-    def register_tesults(self):
-        "Register Tesults with Page"
-        self.tesults_flag = True
-        from utils import Tesults # pylint: disable=import-error,import-outside-toplevel
-        self.tesult_object = Tesults
-
-    def register_browserstack(self):
-        "Register Browser Stack with Page"
-        self.browserstack_flag = True
-        from utils.BrowserStack_Library import BrowserStack_Library # pylint: disable=import-error,import-outside-toplevel
-        self.browserstack_obj = BrowserStack_Library()
 
     def set_calling_module(self,name):
         "Set the test name"
@@ -504,26 +481,10 @@ class Base_Page(Borg, Selenium_Objects, Logging_Objects):
         action_obj.perform()
         self.wait(wait_seconds)
 
-
     def teardown(self):
         "Tears down the driver"
         self.driver.quit()
         self.reset()
-
-
-
-    def report_to_testrail(self,case_id,test_run_id,result_flag,msg=''):
-        "Update Test Rail"
-        if self.testrail_flag is True:
-            msg += '\n'.join(self.msg_list)
-            msg = msg + "\n"
-            if self.browserstack_flag is True:
-                for image in self.image_url_list:
-                    msg += '\n' + '[' + image['name'] + ']('+ image['url']+')'
-                msg += '\n\n' + '[' + 'Watch Replay On BrowserStack' + ']('+ self.session_url+')'
-            self.testrail_object.update_testrail(case_id,test_run_id,result_flag,msg=msg)
-        self.image_url_list = []
-        self.msg_list = []
 
     def add_tesults_case(self, name, desc, suite, result_flag, msg='', files=[], params={}, custom={}):
         "Update Tesults with test results"
